Split PinBallLocationFinder component into two separate components - LocationSearch and NearMeButton for better maintainability and readability also updated SearchForm component
