+++ src/App.css
@@ -1,3 +1,68 @@
-.App {
+* {
+  box-sizing: border-box;
+  font-family: Arial, sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  background-color: #f9f9f9;
+  font-size: 16px;
+  line-height: 1.6;
   text-align: center;
+}
+
+.App {
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+  text-align: left;
+}
+
+
+.location {
+  background-color: #fff;
+  border-radius: 8px;
+  border: 1px solid #ccc;
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.1);
+  padding: 2rem;
+  margin-bottom: 2rem;
+}
+
+.location-name {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+}
+
+.location-address,
+.location-phone {
+  font-size: 1.2rem;
+  margin-bottom: 1rem;
+}
+
+
+.machine-name-label {
+  font-size: 1.2rem;
+  font-weight: bold;
+  margin-top: 1rem;
+  margin-bottom: 0.5rem;
+}
+
+.machine-name {
+  font-size: 1.2rem;
+  margin-left: 1rem;
+  margin-bottom: 1rem;
+}
+
+.machine-name p {
+  margin-bottom: 0.5rem;
+}
+
+
+.search-results {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: flex-start;
 }
