+++ src/components/PinballLocationFinder.jsx
@@ -1,0 +1,14 @@
+import React, { useState } from "react";
+import LocationSearchForm from "./LocationSearchForm.jsx";
+
+
+function PinballLocationFinder() {
+    return (
+        <div className="form-container">
+            <LocationSearchForm />
+            <br />
+        </div>
+
+    );
+}
+export default PinballLocationFinder;
