+++ src/components/SearchResults.jsx
@@ -1,15 +1,20 @@
 import React from "react"
 
-const SearchResults = (locations) => {
-    console.log('locations', locations)
-
+const SearchResults = ({ locations }) => {
     return (
-        locations.locations.map((location) => {
+        locations.map((location) => {
+            console.log(location)
             return (
-                <div key={location.id}>
-                    <p> Name: {location.name}</p>
-                    <p> Adress: {location.street} {location.city}, {location.state} {location.zip} </p>
-                    <p> Phone Number: {location.phone}</p>
+                <div className="location" key={location.id}>
+                    <h3 className="location-name"> Location Name: {location.name}</h3>
+                    <p className="location-adress"> Address: {location.street} {location.city}, {location.state} {location.zip} </p>
+                    {location.phone && <p> Phone Number: {location.phone}</p>}
+                    <p className="machine-name-label">Machine Names: </p>
+                    <div className="machine-name"> {location.machine_names.map(el => {
+                        return (
+                            <p>{el}</p>
+                        )
+                    })}</div>
                 </div>
             )
         })
